SelectPictureBox: New ShowSelection property that supports showing and hiding the selection box.
CVImage:
 * Removed CopyFrom methods.
 * Create is now private.
 * Added CVImage ctor that receives channels and merges them automatically.
 * Added DrawLine methods.
 * Added DrawPixel methods.
 * Some changes in DrawRectangle interface.
 * Added CalcHistogram(binSize) overload.
 * Added CopyRegion which returns a cloned image with RegionOfInterest set to rect.
 * Added CamShift methods which call the OpenCV CAMSHIFT algorithm.
 * Added Clone which returns a new, cloned image.

CVCapture:
 * Added FramesPerSecond property which returns the number of frames per second in the video.
 * Added CreateCompatibleImage method which creates a new CVImage that's compatible with this capture source.

VideoPlayer:
Added a new component to the UI library called VideoPlayer. This component facilitates video processing. It can be used through the designer. You can call Open() to open a video file and then Play or Pause. There are also events that are fired for each new frame and when playing or pausing. Very useful.